re-organized vehicles folder to separate views from controllers
