+++ src/components/ApplicationViews.js
@@ -3,7 +3,7 @@ import { Route, Redirect } from "react-router-dom"
 import DashBoard from "../components/dashboard/DashBoard.js"
 import DealershipList from "./dealership/controllers/searchListContainer"
 import EmployeeList from "./employee/controllers/searchListContainer"
-import VehiclesList from "./vehicle/searchListContainer"
+import VehiclesList from "./vehicle/controllers/searchListContainer"
 import SaleList from "./sale/list"
