+++ src/components/dashboard/DashBoard.js
@@ -4,7 +4,7 @@ import "../../styles/dashboard/dashboard.css"
 import SaleMetrics from "./saleMetrics"
 import Customers from "../customer/list.js";
 import Sales from "../sale/table.js";
-import Vehicles from "../vehicle/dashList.js";
+import Vehicles from "../vehicle/views/dashList.js";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import carnivalImage from "../../images/carnival_cars_image.jpg";
