+++ src/components/vehicle/views/detailsEditWrapper.js
@@ -5,8 +5,8 @@ import Switch from '@material-ui/core/Switch';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormControl from '@material-ui/core/FormControl';
-import { modal } from "../../modules/modal/helpers"
-import SuccessSnackbar from "../modal/snackbar"
+import { modal } from "../../../modules/modal/helpers"
+import SuccessSnackbar from "../../modal/snackbar"
 
 const VehicleDetailsEditWrapper = props => {
     return (
