+++ src/components/vehicle/views/editForm.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { modal } from "../../modules/modal/helpers"
+import { modal } from "../../../modules/modal/helpers"
 
 const VehicleEditForm = props => {
