+++ src/components/vehicle/views/searchList.js
@@ -1,6 +1,6 @@
 import React from "react"
 import VehicleCard from "./vehicleCard"
-import { modal } from "../../modules/modal/helpers"
+import { modal } from "../../../modules/modal/helpers"
 
 const SearchList = props => {
     return (
